+++ assets/js/common/AIAssistant/AgentProgressIndicator/AgentProgressIndicator.jsx
@@ -1,9 +1,18 @@
 import React from 'react';
 import { AuiIf, useAuiState } from '@assistant-ui/react';
 
-import { StatusIndicator } from '../StatusIndicator';
+import Spinner from '@common/Spinner';
 
-export function deriveStatusLabel(content) {
+export function AgentProgressIndicatorView({ label }) {
+  return (
+    <div className="flex items-center gap-2 text-muted-foreground mt-2">
+      <Spinner />
+      <span className="text-sm">{label}</span>
+    </div>
+  );
+}
+
+export function deriveProgressLabel(content) {
   const toolCalls = content.filter((part) => part.type === 'tool-call');
 
   if (toolCalls.length > 0) {
@@ -17,7 +26,7 @@ export function deriveStatusLabel(content) {
   return 'Thinking...';
 }
 
-export function StatusIndicatorContainer() {
+export function AgentProgressIndicator() {
   const message = useAuiState((s) => s.message);
 
   return (
@@ -29,7 +38,9 @@ export function StatusIndicatorContainer() {
         );
       }}
     >
-      <StatusIndicator label={deriveStatusLabel(message.content)} />
+      <AgentProgressIndicatorView
+        label={deriveProgressLabel(message.content)}
+      />
     </AuiIf>
   );
 }
