+++ assets/js/common/AIAssistant/AgentProgressIndicator/index.js
@@ -1,0 +1,1 @@
+export { AgentProgressIndicator } from './AgentProgressIndicator';
