--- assets/js/common/AIAssistant/ConnectionStatusIndicator/index.js
@@ -1,1 +1,0 @@
-export { ConnectionStatusIndicator } from './ConnectionStatusIndicator';
