+++ assets/js/common/AIAssistant/MessageBubble/MessageBubble.jsx
@@ -1,6 +1,15 @@
 import React from 'react';
+import { ErrorPrimitive, MessagePrimitive } from '@assistant-ui/react';
+import { MarkdownTextPrimitive } from '@assistant-ui/react-markdown';
+import '@assistant-ui/react-markdown/styles/dot.css';
+import remarkGfm from 'remark-gfm';
 
-export function MessageBubble({ variant, children }) {
+import { AgentProgressIndicator } from '../AgentProgressIndicator';
+
+const ROOT_CLASS_NAME =
+  'mx-auto w-full max-w-[var(--thread-max-width)] py-2 fade-in slide-in-from-bottom-1 animate-in duration-150';
+
+export function MessageBubbleView({ variant, children }) {
   if (variant === 'user') {
     return (
       <div className="rounded-lg bg-[#e8f5ef] px-5 py-4">
@@ -16,5 +25,47 @@ export function MessageBubble({ variant, children }) {
         {children}
       </div>
     </div>
+  );
+}
+
+function MarkdownText(props) {
+  return (
+    <MarkdownTextPrimitive
+      remarkPlugins={[remarkGfm]}
+      className="aui-md"
+      {...props}
+    />
+  );
+}
+
+function MessageError() {
+  return (
+    <MessagePrimitive.Error>
+      <ErrorPrimitive.Root className="mt-2 rounded-md border border-destructive bg-destructive/10 p-3 text-destructive text-sm dark:bg-destructive/5 dark:text-red-200">
+        <ErrorPrimitive.Message className="line-clamp-2" />
+      </ErrorPrimitive.Root>
+    </MessagePrimitive.Error>
+  );
+}
+
+export function UserMessage() {
+  return (
+    <MessagePrimitive.Root className={ROOT_CLASS_NAME} data-role="user">
+      <MessageBubbleView variant="user">
+        <MessagePrimitive.Parts />
+      </MessageBubbleView>
+    </MessagePrimitive.Root>
+  );
+}
+
+export function AssistantMessage() {
+  return (
+    <MessagePrimitive.Root className={ROOT_CLASS_NAME} data-role="assistant">
+      <MessageBubbleView variant="assistant">
+        <MessagePrimitive.Parts components={{ Text: MarkdownText }} />
+        <MessageError />
+        <AgentProgressIndicator />
+      </MessageBubbleView>
+    </MessagePrimitive.Root>
   );
 }
