--- assets/js/common/AIAssistant/StatusIndicator/index.js
@@ -1,1 +1,0 @@
-export { StatusIndicator } from './StatusIndicator';
