--- assets/js/common/AIAssistant/containers/ChatHeaderContainer.jsx
@@ -1,22 +1,0 @@
-import React, { useCallback } from 'react';
-import { useAui } from '@assistant-ui/react';
-
-import { useAIConnectionStatus } from '../AssistantChatProvider';
-import { ChatHeader } from '../ChatHeader';
-
-export function ChatHeaderContainer({ onClose }) {
-  const aui = useAui();
-  const connectionStatus = useAIConnectionStatus();
-
-  const onNewChat = useCallback(() => {
-    aui.threads().switchToNewThread();
-  }, [aui]);
-
-  return (
-    <ChatHeader
-      connectionStatus={connectionStatus}
-      onNewChat={onNewChat}
-      onClose={onClose}
-    />
-  );
-}
