--- assets/js/common/AIAssistant/containers/ConnectionStatusContainer.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import { useAIConnectionStatus } from '../AssistantChatProvider';
-import { ConnectionStatusIndicator } from '../ConnectionStatusIndicator';
-
-export function ConnectionStatusContainer({ className }) {
-  const status = useAIConnectionStatus();
-  return <ConnectionStatusIndicator status={status} className={className} />;
-}
