+++ assets/js/common/AIAssistant/index.js
@@ -7,6 +7,5 @@ export {
   AssistantChatProvider,
   useAIConnectionStatus,
 } from './AssistantChatProvider';
-export { ConnectionStatusContainer as ConnectionStatusIndicator } from './containers/ConnectionStatusContainer';
 export { WebSocketAIAgent } from './WebSocketAIAgent';
 export { AssistantThread } from './AssistantThread';
